refactor: 🦄  providerにロジックを移行
