+++ src/features/recipeList/lib/index.ts
@@ -1,1 +1,1 @@
-export * from './RecipeList'
+export * from './recipeList'
